Rename Cardano.Wallet.Version to Cardano.Wallet.Application.Version
